feat: introduce Profile model and API controller with CRUD operations.

+++ app/Http/Controllers/Api/ProfileController.php
@@ -12,6 +12,15 @@ use Illuminate\Support\Facades\DB;
 
 class ProfileController extends Controller
 {
+    public function getAllProfiles()
+    {
+        $profile = Profile::all();
+        return response()->json([
+            'status' => Status::Active->value,
+            'message' => 'get success',
+            'profile' => $profile,
+        ]);
+    }
 
     public function store(Request $request)
     {
@@ -20,19 +29,26 @@ class ProfileController extends Controller
             'profile_tax' => 'required',
         ]);
 
+        $profile = Profile::where('profile_id', $request->profile_id)
+            ->orWhere('profile_tax', $request->profile_tax)
+            ->first();
+
+        if ($profile) {
+            return response()->json([
+                'status' => Status::InActive->value,
+                'message' => 'Profile or Tax already exists',
+            ], 200);
+        }
+
         DB::beginTransaction();
 
         try {
             $profile = Profile::create([
                 'profile_id' => $request->profile_id,
                 'profile_tax' => $request->profile_tax,
-                'status' => Status::Active->value,
-                'profile_remark' => $request->profile_remark,
+                'status' => $request->status ?? Status::Active->value,
+                'remark' => $request->profile_remark ?? "",
             ]);
-
-            // $user = User::find(1);
-            // $user->syncPermissions($request->service);
-            // $user->syncPermissions([]);
 
             DB::commit();
 
@@ -58,21 +74,35 @@ class ProfileController extends Controller
             'profile_tax' => 'required',
         ]);
 
+        $profile = Profile::where('profile_id', $request->profile_id)
+            ->Where('profile_tax', $request->profile_tax)
+            ->first();
+
+        if (!$profile) {
+            return response()->json([
+                'status' => Status::InActive->value,
+                'message' => 'Cannot update profile. Please check profile id or profile tax',
+            ], 200);
+        }
+
+        DB::beginTransaction();
         try {
 
-            $profile = Profile::find($request->profile_id);
             $profile->update([
-                'profile_tax' => $request->profile_tax,
-                'profile_remark' => $request->profile_remark,
-                'status' => $request->status,
+                'profile_tax' => isset($request->profile_tax) ? $request->profile_tax : $profile->profile_tax,
+                'remark' => isset($request->profile_remark) ? $request->profile_remark : $profile->remark,
+                'status' => isset($request->status) ? $request->status : $profile->status,
             ]);
-
+            
+            DB::commit();
+            
             return response()->json([
                 'status' => Status::Active->value,
                 'message' => 'update success',
             ]);
 
         } catch(\Exception $e){
+            DB::rollBack();
             return response()->json([
                 'status' => Status::InActive->value,
                 'message' => 'Error updating profile',
@@ -83,11 +113,31 @@ class ProfileController extends Controller
 
     public function destroy(Request $request)
     {
-        $profile = Profile::find($request->profile_id);
-        $profile->delete();
-        return response()->json([
-            'status' => Status::Active->value,
-            'message' => 'delete success'
-        ]);
+        $profile = Profile::where('profile_id', $request->profile_id)
+            ->first();
+
+        if (!$profile) {
+            return response()->json([
+                'status' => Status::InActive->value,
+                'message' => 'Cannot delete profile. Please check profile id or profile tax',
+            ], 200);
+        }
+
+        DB::beginTransaction();
+        try {
+            $profile->delete();
+            DB::commit();
+            return response()->json([
+                'status' => Status::Active->value,
+                'message' => 'delete success'
+            ]);
+        } catch(\Exception $e){
+            DB::rollBack();
+            return response()->json([
+                'status' => Status::InActive->value,
+                'message' => 'Error deleting profile',
+                'description' => $e->getMessage()
+            ], 200);
+        }
     }
 }
